refactor(ai): better file org
